Updated Spatial_BRP code
In this update I added the the double logistic selectivity switch and modified .dat file to include the beta3 and beta4 parameters that are needed for the calculations. I also modified the get_selectivity function to include the ==2 statement with calculations. I cross-checked the cals in R and its all good!