Added recruitment index and adjusted .dat file
In this update I added a recruitment index with lognormal error based on rec_index_sigma. I also updated the .dat file to reflect the population structure we will be using for paper 1....1 population with 3 regions. I wrote the code in R to test a sine function for recruitment with error. I will incorporate that soon as an additional Rec_type switch.  Some other cleaning and stuff too...