Model now accepts maturity varying by area
I made changes to model to allow maturity to vary by area within populations. This involved increasing the dimensions of the maturity parameter. I created another variable that calculates the average maturity across all areas and uses the average for the SPR calculations - will probably need to fix that but the code is now functioning. I did a few tests on the calcs and it looks good but probably need more debug. Still need to add new params to report section.