Changes to vitals calcs
In this update I have added a switch that allows for random allocation of recruitment to areas within populations. It assigns a uniform random deviate (0-1) to each region within years then sums across regions to calculate the proportion of the total recruitment that is assigned to each area. Calculations have been tested and code works. Another change is in by-passing the SPR calcs for the stock-recruitment relationship. The alpha and beta coefficients are now calculated from population steepness, R_ave and unfinished-equilibrium SSB (calculated from initial abundance at age). This,of course, assumes that the model is starting at unfished equilibrium...which might not always be the case...