more additions including survey index
In this addition I have coded a true and observed survey index into the abundance calculations. The true index is spatially-explicit with regional varying selectivity and catchability and can accommodate multiple surveys (as fleets). The calculations summarize the index by population index and also provide regional age composition from the index. The observed survey index includes both observation and process error which is based on specified survey sigma and age-specific aging error. In preparation for manuscript 2, I have coded a matrix that provides the proportion of total abundance in each area in each year based on the survey index...this can be modified to include SSB or any other parameter that might be used to apportion harvest across areas...Again this example used HAKE parameters.