MSY search over multiple folders
This version of MSY_search_spatial will loop over folders in file and conduct the MSY_search with the contents of each folder. Folders in the working directory need to be numbered 1,2,...etc for this to work properly. I have also added the biomass starting values (by region and total) to the csv output for the each run.