Started new tag_estimation model... started removing some switches, added some starter likelihood sections, and made it more efficient by putting most calculated matrices in the data section. Will collaborate with Deroba on how to proceed.
Also added a new folder under Operating Model called "FASTER". This has a new Operating Model that runs twice as fast:

#Original 100 reps
5.440827 mins
#Revised 100 reps
2.513883 mins

Not sure if that's useful, but it gives the same answer as the original. Also in this folder is two new tpl2cpp and tpl2rem .exe files that allow the use of 6d and 7d arrays in the data section which seems like something we may need for the estimation model.