This is a model for the hake high residency example that estimates the optimal Fs to achieve MSY instead of doing the gridsearch. This is achieved by turning on F_switch to 3, F_phase to 3, and adding an objective function loop that maximizes the elements of yield_fleet.
I took out whatever was being doing in the objective function before, but could be put back in (anywhere as it wasn't being used in the minimization process).

This example reproduced (more precisely) the same results as is on the MS1 directory on google drive for this one. I tried it on two other examples, the spatital_random_apportionment1 which seemed to work, and then on the metapopulation 1 which goes of the infinity but it looks like the grid.search would do that too. So we may want some bounds on the F_est.