Made a couple changes to the report section
I made a couple changes to the report section to bring down the dimensions of the observed tag recaptures so I can manipulate in R for export to EM. Also changed the .dat to another tagging design with continued problems with the max_tag_life parameter. It seems that we cannot have a max_tag_life> than the number of tag releases.