This is a working timmy, that I renamed so I didn't overwrite the old one tim_sab... goethels T_pop parameters had a dimension of nps-1, which with a one population model won't work, I changed it to 1, nps because it doesn't matter for sablefish... also found an error of if(y=1) which I changed to if(y==1)... might check OM code for that too.
In addition, the data file needs to have the estimates for surveys and catch in Standard errors, not CVs, and I multiplied tag releases by 1000, and recoveries by 100 and rounded to put them on the right scale and in integers.

This model isn't right yet, but kinda giving reasonable results, but now the gradients are f'ed up. Master Ianelli is looking at it right now, maybe he'll find something brilliant. Anyway, play around with this if you get a chance.

Cheers,
Dana