new wrapper with plots and output
New wrapper for running the OM/EM with plots for diagnostics and outputs. Plotting function generates a pdf with several plots and .txt with outputs of estimated and true values for many parameters. Still need to add the data fit plots (OBS vs TRUE yield and survey index) but coming soon! This is still a simple version for matching OM/EM and single run.