Add features to r_wrapper
Added residual plots for tagging data. Prints out the model structures, if the OM/EM are matching or mismatching, if the model is a diagnostic run (uses true values) and generates a table of error parameters for the OM (similar table for the EM coming soon). This works only for matching but all three population structures. Also reformatted plots to be nicer to review and on standard printing pages. 
Many items presented in tables are reading from the OM dat. Several small but important adjustment made to OM dat to run code successfully.