upload of working metapop .dat
This .dat results in a pretty good estimation for the metapop structure. Overall the model is not great at estimating recruit deviations so maybe it needs to be looked into or it is just a result of this specification. Officially, the matching metapopulation structure with tags is working. This example runs in 16.8 min on my slowish machine.