More features to the wrapper
Now diagnostics wrapper will show the max gradient on the first page of the diagnostic plot. This is read directly from the .par file in the EM folder. I have also added 2 tables at the back of the document, the parameter estimates on the log scale with standard errors (read from the .std) and a correlation table that highlights any correlations with an absolute value > 0.5.