Getting closer to fixing this bastard
Fixed a couple things in the EM to improve simultaneous estimation of init_abund, R_ave, and rec_devs. The model was not putting R_ave in as age 1 of init_abund and the calculations to determine the init_abund of older individuals was incorrectly indexed. Removed the bias correction again on R_ave as it causes bias in the estimation (needs to be further discussed). Fixed some other stuff too, phases and such....but now it is working. Punk.