TIM wrappers needed to run simulations
TIM_diagnostics_SIM is the diagnostics code modified to be sourced from SIMS wrapper to generate diagnostics pdf.  SIM_TIM_editing is the current version of the full sims wrapper. This code will run a user defined number of sims, save and organize the report files and pull important outputs for plotting and summary stats. A number of summary files are generated reporting bias and features of the run. This has been built using only the matching OM/EM so probably more work will be needed for the mismatch.