The master R wrapper...but not done yet!
The SIM_TIM_PLOTS wrapper runs sims, saves data, and generates a number of plots. I have NOT added all the plots I want to yet, but they are coming. The user specifies a number of inputs at the beginning of the code (eg run sims, how many, save values, make plots, run diag code, etc) and the code does the rest. The best part...this WORKS with the mismatch.